Failed to refactor code
Goal was to remove getLanguage function from addPronunciation. Spent a
lot of time and couldn't get it to work. Committing so you can see what
I tried. And then will revert.
